+++ doc/changelog.md
@@ -2,6 +2,8 @@
 
 ## 0.0.3 (Alpha)
 * Location recording
+* Fix for faulty automatic synchronisation
+* Synchronisation as foreground async (with notification)
 
 ## 0.0.2 (Alpha)
 * Show risk summary